+++ client/constants/flags.ts
@@ -1,0 +1,3 @@
+import env from '@beam-australia/react-env';
+
+export const FLAG_DISABLE_SIGNUPS = env('FLAG_DISABLE_SIGNUPS') === 'true';